requirements.txt
python_dotenv
matplotlib
requests